moved navigation bar icon container in containers/
