+++ src/components/containers/navigation-bar-icon-container.tsx
