+++ src/components/icons/key-icon/key-icon.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { NavigationBarIconContainer } from "../../navigation-bar-icon-container";
+import { NavigationBarIconContainer } from "../../containers/navigation-bar-icon-container";
 
 export const KeyIcon: React.FC = () => {
   return (
