+++ src/components/icons/world-icon/black-world-icon/black-world-icon.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { NavigationBarIconContainer } from "../../../navigation-bar-icon-container";
+import { NavigationBarIconContainer } from "../../../containers/navigation-bar-icon-container";
 
 export const BlackWorldIcon: React.FC = () => {
   return (
